+++ hutool-core/src/main/java/cn/hutool/core/convert/impl/ReferenceConverter.java
@@ -37,7 +37,7 @@ public class ReferenceConverter extends AbstractConverter<Reference> {
 		//尝试将值转换为Reference泛型的类型
 		Object targetValue = null;
 		final Type paramType = TypeUtil.getTypeArgument(targetType);
-		if(false == TypeUtil.isUnknow(paramType)){
+		if(false == TypeUtil.isUnknown(paramType)){
 			targetValue = ConverterRegistry.getInstance().convert(paramType, value);
 		}
 		if(null == targetValue){