+++ hutool-core/src/main/java/cn/hutool/core/lang/reflect/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 提供反射相关功能对象和类
+ *
+ * @author looly
+ * @since 5.4.2
+ */
+package cn.hutool.core.lang.reflect;